Conflicts:
	core/Makefile
	core/version_defaults.mk
	tools/device/makerecoveries.sh

+++ tools/device/mkvendor.sh
@@ -3,11 +3,13 @@
 function usage
 {
     echo Usage:
-    echo "  $(basename $0) manufacturer device boot.img"
+    echo "  $(basename $0) manufacturer device [boot.img]"
     echo "  The boot.img argument is the extracted recovery or boot image."
+    echo "  The boot.img argument should not be provided for devices"
+    echo "  that have non standard boot images (ie, Samsung)."
     echo
     echo Example:
-    echo "  $(basename $0) ~/Downloads/recovery-passion.img motorola sholes"
+    echo "  $(basename $0) motorola sholes ~/Downloads/recovery-sholes.img"
     exit 0
 }
 
@@ -27,19 +29,6 @@ then
     usage
 fi
 
-if [ -z "$BOOTIMAGE" ]
-then
-    usage
-fi
-
-if [ -z "$UNPACKBOOTIMG" ]
-then
-    echo unpackbootimg not found. Is your android build environment set up and have the host tools been built?
-    exit 0
-fi
-
-BOOTIMAGEFILE=$(basename $BOOTIMAGE)
-
 ANDROID_TOP=$(dirname $0)/../../../
 pushd $ANDROID_TOP > /dev/null
 ANDROID_TOP=$(pwd)
@@ -51,22 +40,46 @@ TEMPLATE_DIR=$(pwd)
 popd > /dev/null
 
 DEVICE_DIR=$ANDROID_TOP/device/$MANUFACTURER/$DEVICE
-echo Output will be in $DEVICE_DIR
-mkdir -p $DEVICE_DIR
 
-TMPDIR=/tmp/bootimg
-rm -rf $TMPDIR
-mkdir -p $TMPDIR
-cp $BOOTIMAGE $TMPDIR
-pushd $TMPDIR > /dev/null
-unpackbootimg -i $BOOTIMAGEFILE > /dev/null
-BASE=$(cat $TMPDIR/$BOOTIMAGEFILE-base)
-CMDLINE=$(cat $TMPDIR/$BOOTIMAGEFILE-cmdline)
-PAGESIZE=$(cat $TMPDIR/$BOOTIMAGEFILE-pagesize)
-export SEDCMD="s#__CMDLINE__#$CMDLINE#g"
-echo $SEDCMD > $TMPDIR/sedcommand
-cp $TMPDIR/$BOOTIMAGEFILE-zImage $DEVICE_DIR/kernel
-popd > /dev/null
+if [ ! -z "$BOOTIMAGE" ]
+then
+    if [ -z "$UNPACKBOOTIMG" ]
+    then
+        echo unpackbootimg not found. Is your android build environment set up and have the host tools been built?
+        exit 0
+    fi
+
+    BOOTIMAGEFILE=$(basename $BOOTIMAGE)
+
+    echo Output will be in $DEVICE_DIR
+    mkdir -p $DEVICE_DIR
+
+    TMPDIR=/tmp/bootimg
+    rm -rf $TMPDIR
+    mkdir -p $TMPDIR
+    cp $BOOTIMAGE $TMPDIR
+    pushd $TMPDIR > /dev/null
+    unpackbootimg -i $BOOTIMAGEFILE > /dev/null
+    mkdir ramdisk
+    pushd ramdisk > /dev/null
+    gunzip -c ../$BOOTIMAGEFILE-ramdisk.gz | cpio -i
+    popd > /dev/null
+    BASE=$(cat $TMPDIR/$BOOTIMAGEFILE-base)
+    CMDLINE=$(cat $TMPDIR/$BOOTIMAGEFILE-cmdline)
+    PAGESIZE=$(cat $TMPDIR/$BOOTIMAGEFILE-pagesize)
+    export SEDCMD="s#__CMDLINE__#$CMDLINE#g"
+    echo $SEDCMD > $TMPDIR/sedcommand
+    cp $TMPDIR/$BOOTIMAGEFILE-zImage $DEVICE_DIR/kernel
+    popd > /dev/null
+else
+    mkdir -p $DEVICE_DIR
+    touch $DEVICE_DIR/kernel
+    BASE=10000000
+    CMDLINE=no_console_suspend
+    PAGESIZE=00000800
+    export SEDCMD="s#__CMDLINE__#$CMDLINE#g"
+    echo $SEDCMD > $TMPDIR/sedcommand
+fi
 
 for file in $(find $TEMPLATE_DIR -name '*.template')
 do
@@ -74,12 +87,18 @@ do
     cat $file | sed s/__DEVICE__/$DEVICE/g | sed s/__MANUFACTURER__/$MANUFACTURER/g | sed -f $TMPDIR/sedcommand | sed s/__BASE__/$BASE/g | sed s/__PAGE_SIZE__/$PAGESIZE/g > $OUTPUT_FILE
 done
 
-mv $DEVICE_DIR/device.mk $DEVICE_DIR/device_$DEVICE.mk
-RECOVERY_FSTAB=$TMPDIR/ramdisk/etc/recovery.fstab
-if [ -f "$RECOVERY_FSTAB" ]
+if [ ! -z "$TMPDIR" ]
 then
-  cp $RECOVERY_FSTAB $DEVICE_DIR/recovery.fstab
+    RECOVERY_FSTAB=$TMPDIR/ramdisk/etc/recovery.fstab
+    if [ -f "$RECOVERY_FSTAB" ]
+    then
+        cp $RECOVERY_FSTAB $DEVICE_DIR/recovery.fstab
+    fi
 fi
+
+
+mv $DEVICE_DIR/device.mk $DEVICE_DIR/device_$DEVICE.mk
+
 
 echo Done!
 echo Use the following command to set up your build environment:
